Add generic error handler
This will automatically reformat HttpError (custom error type) errors as JSON
responses with an error object as the payload. Note: this error object is not
wrapped in another object.

The status of the response will correspond to error.code. It will also return a
generic 500 error if passed a generic Error type.