Silence Mongo deprecation warnings
Unfortunately, there is a bug with the newUrlParser at the moment, and it does
not correctly set the default database port. As a result, the DATABASE_URLs
have been updated to include the default port.

Additionally, Model.count() is now deprecated in favor of countDocuments, so
this was updated in the users router.