Fix 500 error when saving a missing aisleLocation
This would prevent checking off an item that did not have aisleLocation
information embedded in it. Previously, POST would not create an
aisleLocation for every item. Thus, old item entries would cause the server to
throw a typeerror. Now, old items will be updated to have aisle information.

Additionally, we now validate against the situation where a user attempts to
enter aisle information for a list that is not attached to a store (we throw a
HTTP 422 error).